Separate HTTP request parsing into its own channel handler
